eld_v2/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
telethon==1.36.0
aiohttp==3.10.5
python-dateutil==2.9.0
python-dotenv==1.0.1
aiosqlite==0.20.0
